% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap.pensynth}
\alias{bootstrap.cvpensynth}
\title{Bayesian bootstrap inference}
\usage{
bootstrap(object, n_iter = 1000L, cl)

\method{bootstrap}{pensynth}(object, n_iter = 1000L, cl)

\method{bootstrap}{cvpensynth}(object, n_iter = 1000L)
}
\arguments{
\item{object}{a fitted \code{pensynth} or \code{cvpensynth} object}

\item{n_iter}{\code{int} number of bootstrap iterations (default 1000)}
}
\description{
Produce Bayesian bootstrap prediction intervals and ATE intervals
for pensynth objects. This is done through reweighing the columns
of the data (X0 and X1) using Dirichlet weights.
}
