% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_pensynth.R
\name{get_mse_cvpensynth}
\alias{get_mse_cvpensynth}
\title{Get average mse for cvpensynth object}
\usage{
get_mse_cvpensynth(x)
}
\arguments{
\item{fit}{The cvpensynth object}
}
\value{
numeric
}
\description{
Get average mse for cvpensynth object
}
\keyword{internal}
