% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convex_hull.R
\name{in_convex_hull}
\alias{in_convex_hull}
\title{Is the treated unit in the convex hull}
\usage{
in_convex_hull(X1, X0)
}
\arguments{
\item{X1}{\verb{N_covars by 1 matrix} of treated unit covariates}

\item{X0}{\verb{N_covars by N_donors matrix} of donor unit covariates}
}
\value{
\code{bool} whether the treated unit is in the convex hull of
the donor units.
}
\description{
This function finds out if the treated unit is in the convex hull
of the donor units.
}
\seealso{
\code{\link[geometry:convhulln]{geometry::convhulln()}} \code{\link[geometry:inhulln]{geometry::inhulln()}}
}
