% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{mvrarnorm}
\alias{mvrarnorm}
\title{Generate data from independent multivariate normal
distribution with autocorrelation parameter}
\usage{
mvrarnorm(n, p, mean = 0, sd = 1, phi = 0)
}
\arguments{
\item{n}{number of observations.}

\item{mean}{marginal mean}

\item{sd}{marginal standard deviation}

\item{phi}{autocorrelation parameter (-1 < phi < 1)}
}
\value{
matrix of numeric values
}
\description{
Generate data from independent multivariate normal
distribution with autocorrelation parameter
}
\details{
Note that, unlike \code{\link[stats:Normal]{stats::rnorm()}}, this function is
not vectorized over mean, sd, or phi.
}
\keyword{internal}
