% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placebo_test.R
\name{placebo_test}
\alias{placebo_test}
\alias{placebo_test.pensynth}
\alias{placebo_test.cvpensynth}
\title{Placebo permutation test for pensynth}
\usage{
placebo_test(object, Y1, Y0)

\method{placebo_test}{pensynth}(object, Y1, Y0)

\method{placebo_test}{cvpensynth}(object, Y1, Y0)
}
\arguments{
\item{object}{a fitted \code{pensynth} or \code{cvpensynth} object}

\item{Y1}{the post-intervention outcome of the treated unit}

\item{Y0}{the post-intervention outcome of the donor units
(with N_donors columns)}
}
\value{
A list with two elements
\itemize{
\item E1, the treated unit effect, computed as \code{Y1 - Y0 \%*\% w}
\item E0, the donor unit effects, computed in the same way but
using the permutation test's weights.
\item ATE1, the estimated ATE of the treated unit
\item ATE0, the estimated ATE of the donor units
}
}
\description{
Perform a permutation test on a pensynth object, in the sense
of Abadie, Diamond, and Hainmueller (2010). The pensynth
method is performed multiple times, treating each donor as the
treated unit and the treated unit with the remaining donors as
the donor units.
}
\references{
Abadie, A., Diamond, A., & Hainmueller, J. (2010).
Synthetic control methods for comparative case studies:
Estimating the effect of California’s tobacco control program.
Journal of the American statistical Association, 105(490),
493-505.
}
