% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{plot.pensynthtest}
\alias{plot.pensynthtest}
\title{Plotting a pensynth permutation object}
\usage{
\method{plot}{pensynthtest}(object, ...)
}
\arguments{
\item{object}{a \code{pensynthtest} object}

\item{...}{additional parameters passed to \code{plot}}
}
\description{
Plotting the reference distribution and the
estimated treatement effect for the treated unit
for the pensynth permutation test.
}
\seealso{
\code{\link[base:plot]{base::plot()}}
}
