% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_pensynth.R
\name{plot_path}
\alias{plot_path}
\title{Plotting for cross-validated penalized synthetic control objects}
\usage{
plot_path(object, ...)
}
\arguments{
\item{object}{a \code{cvpensynth} output object}

\item{...}{additional arguments passed to \code{plot()}}
}
\value{
No return value, called for side effects
}
\description{
Displays a mean squared error curve and weights curve as a function
of lambda, the penalization parameter.
}
\seealso{
\code{\link[=cv_pensynth]{cv_pensynth()}} \code{\link[=pensynth]{pensynth()}}
}
