% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensynth.R
\name{predict.pensynth}
\alias{predict.pensynth}
\title{Create prediction from pensynth model}
\usage{
\method{predict}{pensynth}(object, newdata, ...)
}
\arguments{
\item{object}{a fitted cvpensynth model}

\item{newdata}{N_values * N_donors matrix of
values for the donor units.}

\item{...}{ignored}
}
\value{
a matrix (column vector) of predicted values
}
\description{
Matrix multiplies the values in \code{newdata} by the unit weights
extracted from the pensynth object to produce predicted
values.
}
