% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_pensynth.R
\name{print.cvpensynth}
\alias{print.cvpensynth}
\title{Print cvpensynth model}
\usage{
\method{print}{cvpensynth}(x, ...)
}
\arguments{
\item{x}{a cvpensynth object}

\item{...}{ignored}
}
\description{
Print cvpensynth model
}
