% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensynth.R
\name{print.pensynth}
\alias{print.pensynth}
\title{Print pensynth model}
\usage{
\method{print}{pensynth}(x, ...)
}
\arguments{
\item{x}{a pensynth object}

\item{...}{ignored}
}
\value{
the pensynth object, invisibly
}
\description{
Print pensynth model
}
