% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{standardize_X}
\alias{standardize_X}
\title{Function to scale the X matrices in synthetic control}
\usage{
standardize_X(X1, X0)
}
\arguments{
\item{X1}{the X1 matrix}

\item{X0}{the X0 matrix}
}
\description{
The default implementation of synthetic control first
scales the X matrices using the mean and standard
deviation of the X1 and X0 matrices together.
}
\details{
See the pre-processing in \code{\link[Synth:synth]{Synth::synth()}}.
}
\seealso{
\code{\link[base:scale]{base::scale()}}
}
\keyword{internal}
