% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_pensynth.R
\name{predict.cvpensynth}
\alias{predict.cvpensynth}
\title{Create prediction from pensynth model}
\usage{
\method{predict}{cvpensynth}(object, newdata, lambda, ...)
}
\arguments{
\item{object}{a fitted pensynth model}

\item{newdata}{N_values * N_donors matrix of
values for the donor units.}

\item{lambda}{desired lambda value}
}
\description{
Create prediction from pensynth model
}
